+++ CHANGELOG.md
@@ -6,6 +6,7 @@
   - bottom horizontal scrollbar not responding on small screens
   - remote file browser: hitting a letter to select a file based on its name was looping on the first two matches only
   - editor: preserve backslashes in enquoted strings (not in objects/arrays)
+  - editor: clear the editor when selected widget is removed by an undo/redo action
 - widgets
   - `logScale` can be set to a number to use a logarithm base different from the default
   - knob: `compact` mode is back